ref: rename beans package
