+++ src/main/java/com/esosa/pass_manager_hexagonal/infrastructure/beans/UserBeans.java
@@ -1,4 +1,4 @@
-package com.esosa.pass_manager_hexagonal.infrastructure.config;
+package com.esosa.pass_manager_hexagonal.infrastructure.beans;
 
 import com.esosa.pass_manager_hexagonal.application.services.UserService;
 import com.esosa.pass_manager_hexagonal.application.usecases.user.GetUserUserCaseImpl;
